Move event components into corresponding folder
