+++ src/app/manage/events/[id]/page.jsx
@@ -18,10 +18,10 @@ import ActionPalette from "components/ActionPalette";
 
 import EventDetails from "components/events/EventDetails";
 import EventBudget from "components/events/EventBudget";
-import EventBillStatus from "components/events/EventBillStatus";
-import EventReportStatus from "components/events/EventReportStatus";
+import EventBillStatus from "components/events/bills/EventBillStatus";
+import EventReportStatus from "components/events/report/EventReportStatus";
 import EventApprovalStatus from "components/events/EventApprovalStatus";
-import { DownloadEvent } from "components/events/EventpdfDownloads";
+import { DownloadEvent } from "components/events/report/EventpdfDownloads";
 import {
   EditEvent,
   CopyEvent,
@@ -63,11 +63,13 @@ export default async function ManageEventID({ params }) {
     eventid: id,
   });
 
-  const { data: { activeClubs }, } = await getClient().query(GET_ACTIVE_CLUBS);
+  const {
+    data: { activeClubs },
+  } = await getClient().query(GET_ACTIVE_CLUBS);
 
   const { data: { userMeta, userProfile } = {} } = await getClient().query(
     GET_USER,
-    { userInput: null },
+    { userInput: null }
   );
   const user = { ...userMeta, ...userProfile };
 
@@ -91,7 +93,13 @@ export default async function ManageEventID({ params }) {
             { status: event?.status, location: event?.location },
           ]}
           right={getActions(event, { ...userMeta, ...userProfile })}
-          downloadbtn={<DownloadEvent event={event} clubs={activeClubs} pocProfile={pocProfile} />}
+          downloadbtn={
+            <DownloadEvent
+              event={event}
+              clubs={activeClubs}
+              pocProfile={pocProfile}
+            />
+          }
         />
         <EventDetails showCode event={event} />
         <Divider sx={{ borderStyle: "dashed", my: 2 }} />
